% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectordata.R
\name{vectordata}
\alias{vectordata}
\title{VectorData}
\usage{
vectordata(x)
}
\arguments{
\item{x}{a vector of any length}
}
\value{
Median, Range, Length, SquaredVector, Cubedvector
}
\description{
Takes a vector and returns the following values of the vector
}
